+++ components/language-picker/common/find-language/langtags.txt
