+++ components/language-picker/common/find-language/regionsAndScripts.ts
