+++ components/language-picker/common/find-language/vitest.config.ts
@@ -2,9 +2,9 @@
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  test: {
-    expect: {
-      requireAssertions: true,
-    },
-  },
+  // test: {
+  //   expect: {
+  //     requireAssertions: true,
+  //   },
+  // },
 });
